fix: improve FTA s3-filesystem-storage
