+++ src/service/s3/storage/filesystem/s3-filesystem-safety.iso.test.ts
@@ -9,8 +9,8 @@ import {
   assertThrowsErrorAsync,
 } from "@kensio/smartass";
 import { FilesystemS3BucketStorage } from "./s3-filesystem-storage.js";
-import { SimS3Object } from "../object/s3-object.js";
-import { makeTempDir } from "../../../util/filesystem/temp-dir.js";
+import { SimS3Object } from "../../object/s3-object.js";
+import { makeTempDir } from "../../../../util/filesystem/temp-dir.js";
 
 describe("Filesystem simulated S3 storage safety", () => {
   it("rejects relative storage directory path", () => {
