+++ src/service/s3/storage/filesystem/s3-filesystem-storage.ts
@@ -1,8 +1,10 @@
-import type { SimS3BucketStorage } from "./s3-bucket-storage.js";
-import { SimS3Object, SimS3ObjectMetadata } from "../object/s3-object.js";
+import type { SimS3BucketStorage } from "../s3-bucket-storage.js";
+import { SimS3Object } from "../../object/s3-object.js";
 import path from "node:path";
-import { mkdir, readdir, readFile, stat, writeFile } from "node:fs/promises";
+import { mkdir, readdir, readFile, writeFile } from "node:fs/promises";
 import { FilesystemS3StorageSafety } from "./s3-filesystem-safety.js";
+import { metadataForFilesystemS3ObjectKey } from "./s3-filesystem-object-metadata.js";
+import { filesystemPathExists } from "./filesystem-path-exists.js";
 
 interface FilesystemS3BucketStorageProps {
   readonly directoryPath: string;
@@ -46,7 +48,7 @@ export class FilesystemS3BucketStorage implements SimS3BucketStorage {
       return new SimS3Object({
         key,
         body,
-        metadata: this.metadataForObjectKey(key),
+        metadata: metadataForFilesystemS3ObjectKey(key),
       });
     } catch (error) {
       if (
@@ -124,20 +126,8 @@ export class FilesystemS3BucketStorage implements SimS3BucketStorage {
   }
 
   private async listObjectKeys(): Promise<string[]> {
-    try {
-      // eslint-disable-next-line security/detect-non-literal-fs-filename
-      await stat(this.directoryPath);
-    } catch (error) {
-      if (
-        error instanceof Error &&
-        "code" in error &&
-        error.code === "ENOENT"
-      ) {
-        return [];
-      }
-
-      /* v8 ignore next */
-      throw error;
+    if (!(await filesystemPathExists(this.directoryPath))) {
+      return [];
     }
 
     return await this.listObjectKeysInDirectory(this.directoryPath);
@@ -175,55 +165,4 @@ export class FilesystemS3BucketStorage implements SimS3BucketStorage {
 
     return [...keys, ...nestedKeys.flat()];
   }
-
-  /**
-   * Make up reasonable metadata for an Object based on the file on disk.
-   */
-  private metadataForObjectKey(key: string): SimS3ObjectMetadata {
-    const contentType = this.contentTypeForObjectKey(key);
-
-    /* v8 ignore if -- defensive fallback */
-    if (contentType === undefined) {
-      return new SimS3ObjectMetadata();
-    }
-
-    return new SimS3ObjectMetadata({
-      "content-type": contentType,
-    });
-  }
-
-  /**
-   * Guess a reasonable content type for an Object key based on its file
-   * extension on disk.
-   */
-  private contentTypeForObjectKey(key: string): string | undefined {
-    const extension = path.extname(key).toLowerCase();
-
-    return contentTypesByExtension.get(extension);
-  }
 }
-
-const contentTypesByExtension: ReadonlyMap<string, string> = new Map([
-  [".css", "text/css"],
-  [".eot", "application/vnd.ms-fontobject"],
-  [".gif", "image/gif"],
-  [".html", "text/html"],
-  [".htm", "text/html"],
-  [".ico", "image/x-icon"],
-  [".jpeg", "image/jpeg"],
-  [".jpg", "image/jpeg"],
-  [".js", "text/javascript"],
-  [".mjs", "text/javascript"],
-  [".json", "application/json"],
-  [".map", "application/json"],
-  [".png", "image/png"],
-  [".otf", "font/otf"],
-  [".svg", "image/svg+xml"],
-  [".ttc", "font/collection"],
-  [".ttf", "font/ttf"],
-  [".txt", "text/plain"],
-  [".webp", "image/webp"],
-  [".woff", "font/woff"],
-  [".woff2", "font/woff2"],
-  [".xml", "application/xml"],
-]);
